pamp_app/permissions.py
from rest_framework.permissions import BasePermission, SAFE_METHODS


class IsOwnerOrReadOnly(BasePermission):
    """Allow safe methods to everyone and write methods only to owners."""

    def has_object_permission(self, request, view, obj):
        if request.method in SAFE_METHODS:
            return True

        return getattr(obj.profile, 'user_id', None) == getattr(request.user, 'id', None)
