requirements.txt
# Django and DRF
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
asgiref==3.8.1
sqlparse==0.5.3

# Database and cache
psycopg[binary]==3.2.6
redis==5.2.1
celery==5.5.1
pillow==11.2.1
django-storages[s3]==1.14.6
boto3==1.37.38

# Typing
mypy==1.15.0
django-stubs==5.2.6
djangorestframework-stubs==3.16.3
types-requests==2.32.0.20250306

# Authentication
dj-rest-auth==5.0.1
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4

# Security
cffi==1.17.1
cryptography==43.0.1
defusedxml==0.7.1
django-extensions==3.2.3
markupsafe==2.1.5
oauthlib==3.2.2
pycparser==2.22
pyjwt==2.9.0
pyopenssl==24.2.1
python3-openid==3.2.0
requests-oauthlib==2.0.0
werkzeug==3.0.4

# Telegram Bot
requests
python-telegram-bot==21.10
pytz==2024.2
python-dotenv==1.0.1
httpx==0.28.1
djangorestframework-api-key==3.0.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2026.3.1
